requirements.txt
# SHSE — Python Dependencies

# Web Framework
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0

# Async Task Queue
Celery==5.3.4
redis==5.0.1
celery-redbeat==2.3.3

# OpenSearch
opensearch-py==2.4.2

# Ollama Integration
requests==2.31.0

# Authentication
bcrypt==4.1.2
Authlib==1.3.0
pyotp==2.9.0
webauthn==2.2.0

# YAML Configuration
PyYAML==6.0.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Token counting (embedding context-window guard)
tiktoken==0.12.0

# CPU-based embedding fallback (semantic rail when LLM API is down)
sentence-transformers>=3.0.0

# MCP Server
fastapi>=0.115
uvicorn[standard]>=0.30
mcp>=1.0

# Development/Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-playwright>=0.4.0
